feat: UserService를 인터페이스로 변경 후 구현 클래스 추가
