+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,41 @@
+package com.techcourse.service;
+
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+import java.sql.Connection;
+import javax.sql.DataSource;
+
+public class AppUserService implements UserService {
+
+    private final DataSource dataSource;
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+
+    public AppUserService(DataSource dataSource, UserDao userDao, UserHistoryDao userHistoryDao) {
+        this.dataSource = dataSource;
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        Connection connection = TransactionSynchronizationManager.getResource(dataSource);
+        final var user = findById(id);
+        user.changePassword(newPassword);
+        userDao.update(connection, user);
+        userHistoryDao.log(connection, new UserHistory(user, createBy));
+    }
+}
