+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,50 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TxUserService implements UserService {
+
+    private final DataSource dataSource;
+    private final UserService userService;
+
+    public TxUserService(DataSource dataSource, UserService userService) {
+        this.dataSource = dataSource;
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userService.save(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createdBy) {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+            userService.changePassword(id, newPassword, createdBy);
+            connection.commit();
+        } catch (Exception exception) {
+            try {
+                connection.rollback();
+            } catch (SQLException sqlException) {
+                throw new DataAccessException("DB 처리에 실패하여 DB 롤백을 시도하던 도중 오류가 발생하였습니다.", sqlException);
+            }
+            throw new DataAccessException("DB 처리 도중 오류가 발생하였습니다.", exception);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+            TransactionSynchronizationManager.unbindResource(dataSource);
+        }
+    }
+}
